+++ src/cpp/include/tbsla/cpp/Matrix.hpp
@@ -11,24 +11,24 @@ namespace tbsla { namespace cpp {
 class Matrix {
   public:
     friend std::ostream & operator<<( std::ostream &os, const Matrix &m) { return m.print(os); };
-    virtual double* spmv(const double* v, int vect_incr = 0) const = 0;
-    virtual inline void Ax(double* r, const double* v, int vect_incr = 0) const = 0;
+    virtual double* spmv(const double* v, long long int vect_incr = 0) const = 0;
+    virtual inline void Ax(double* r, const double* v, long long int vect_incr = 0) const = 0;
 	void make_stochastic(double* s);
-    void AAxpAx(double* r, double* v, double* buffer, int vect_incr = 0) const;
-    void AAxpAxpx(double* r, double* v, double* buffer, int vect_incr = 0) const;
-    double* a_axpx_(const double* v, int vect_incr = 0) const;
+    void AAxpAx(double* r, double* v, double* buffer, long long int vect_incr = 0) const;
+    void AAxpAxpx(double* r, double* v, double* buffer, long long int vect_incr = 0) const;
+    double* a_axpx_(const double* v, long long int vect_incr = 0) const;
     double* & saxpy(const double* x, double* y);
-    int get_n_row() const {return n_row;}
-    int get_n_col() const {return n_col;}
-    int get_f_row() const {return f_row;}
-    int get_f_col() const {return f_col;}
-    int get_ln_row() const {return ln_row;}
-    int get_ln_col() const {return ln_col;}
-    int get_pr() const {return pr;}
-    int get_pc() const {return pc;}
-    int get_NR() const {return NR;}
-    int get_NC() const {return NC;}
-    long int get_nnz() const {return nnz;};
+    long long int get_n_row() const {return n_row;}
+    long long int get_n_col() const {return n_col;}
+    long long int get_f_row() const {return f_row;}
+    long long int get_f_col() const {return f_col;}
+    long long int get_ln_row() const {return ln_row;}
+    long long int get_ln_col() const {return ln_col;}
+    long long int get_pr() const {return pr;}
+    long long int get_pc() const {return pc;}
+    long long int get_NR() const {return NR;}
+    long long int get_NC() const {return NC;}
+    long long int get_nnz() const {return nnz;};
     std::string get_vectorization() const {return "None";}
 
     virtual std::ostream & write(std::ostream &os) = 0;
@@ -38,11 +38,11 @@ class Matrix {
 
     virtual void NUMAinit() = 0;
 
-    virtual void fill_cdiag(int n_row, int n_col, int cdiag, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
-    virtual void fill_cqmat(int n_row, int n_col, int c, double q, unsigned int seed_mult = 1, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
-    virtual void fill_random(int n_row, int n_col, double nnz_ratio, unsigned int seed_mult = 1, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
-    virtual void fill_brain(int n_row, int n_col, int* neuron_type, std::vector<std::vector<double> > proba_conn, std::vector<std::unordered_map<int,std::vector<int> > > brain_struct, unsigned int seed_mult, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
-    virtual void fill_cdistrib(int n_row, int n_col, int nnz, int pr = 0, int pc = 0, int NR = 1, int NC = 1) = 0;
+    virtual void fill_cdiag(long long int n_row, long long int n_col, long long int cdiag, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1) = 0;
+    virtual void fill_cqmat(long long int n_row, long long int n_col, long long int c, double q, unsigned long long int seed_mult = 1, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1) = 0;
+    virtual void fill_random(long long int n_row, long long int n_col, double nnz_ratio, unsigned long long int seed_mult = 1, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1) = 0;
+    virtual void fill_brain(long long int n_row, long long int n_col, int* neuron_type, std::vector<std::vector<double> > proba_conn, std::vector<std::unordered_map<int,std::vector<int> > > brain_struct, unsigned long long int seed_mult, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1) = 0;
+    virtual void fill_cdistrib(long long int n_row, long long int n_col, long long int nnz, long long int pr = 0, long long int pc = 0, long long int NR = 1, long long int NC = 1) = 0;
 	
 	virtual void get_row_sums(double* buffer) = 0;
 	virtual void normalize_rows(double* s) = 0;
@@ -51,8 +51,8 @@ class Matrix {
 	virtual void set_diag(double* s) = 0;
 
   protected:
-    int n_row, n_col, f_row, f_col, ln_row, ln_col, pr, pc, NR, NC;
-    long int nnz;
+    long long int n_row, n_col, f_row, f_col, ln_row, ln_col, pr, pc, NR, NC;
+    long long int nnz;
 
 };
