+++ src/cpp/include/tbsla/cpp/utils/range.hpp
@@ -5,8 +5,8 @@
 
 namespace tbsla { namespace utils { namespace range {
 
-  size_t lnv(size_t size, int local_pos, int number_pos);
-  size_t pflv(size_t size, int local_pos, int number_pos);
+  size_t lnv(size_t size, long long int local_pos, long long int number_pos);
+  size_t pflv(size_t size, long long int local_pos, long long int number_pos);
 }}}
 
 #endif
